Update ecr repo push command to use get-login-password (#1278)
This change updates the push command shown in the ECR Repo viewer. The command referenced an AWS CLI command (`aws ecr get-login`) which no longer exists. The command has been updated to use `get-login-password`.  The old command output a docker login command and password, and the new command only outputs the password, so this change also provides users with a command that appropriate pipes the output in order to successfully log in.

This change is a fix for: https://github.com/aws/aws-toolkit-visual-studio/issues/246