Add v10 folder for 2022 Lambda blueprints (#1280)
Adds a v10 folder for Lambda blueprints targeted only for VS2022. It currently contains a copy of VS2022 blueprints from the v9 folder.
An if-def has been added such that 2019 or lower versions of VS would continue to target the v9 folder. This prevents the unnecessary increase in extension size by updating only required(currently supported) versions of VS.