Add error messages when CAWS API calls fail (#1299)
This change adds validation messages to Spaces/Projects/Repositories/Path (path doesn't have validation code to generate a validation message yet, just the control added for an upcoming PR that adds the validation code).

Added RelayValueConverter while we take a look at our existing value converters and streamline them. This allows keeping the actual conversion code near where it is used and reduces blast radius if the behavior needs to be changed as it will only impact the local view/control rather than multiple across the toolkit.

Added ValidationErrorsBox control which can be bound to a control (via SourceElement) that may have validation errors. This control is meant to reduce the amount of boilerplate XAML needed wherever validation errors need to be displayed. It also allows a RetryCommand to be bound to execute retry functionality (if it is not bound, the Retry hyperlink is not displayed). There is also an Icon property that allows the icon to display on the left of the message to be bound as well.

The CloneCodeCatalystRepositoryDialog classes were updated to support validation messages. 3 RetryCommands were added for Space/Project/Repository. The Refresh... methods in the view model were updated to use the BaseModel.DataErrorInfo to add errors when the CodeCatalyst API calls fail.