Add basic LocalPath rewriting when user is just clicking through repositories. (#1303)
With this update, the Path in the CloneCodeCatalystRepository dialog behaves like other git UIs in VS.

The Git default location is taken from Tools/Options if set, otherwise the default (which is in the VS code) of %USERPROFILE%\Source\Repos is used. (This will be back ported to CodeCommit IDE-8847)
The selected repository name is automatically appended to the path if the previous repository name (or "" on no prior selection) is at the end of the path. The previous repo name is removed first. This basically let's the user get some automatic name handling, but if they start changing the path in the textbox on their own, we don't attempt to rewrite it.
Migrated Git to IGetService/GetService, but currently GitService just instantiated and threaded through the CloneCodeCatalyst* classes until we sort out our service provider story. This will allow git helper functionality to be available in assemblies that don't take a dependency on the VSSDK. As part of this, no longer passing IProgress to CloneAsync, as ServiceProgressData is a VSSDK type. The IProgress never worked write with the VSSDK CloneAsync method anyway.

Added more unit tests around LocalPath for new behavior.

While the current path rewrite behavior is very basic, it will hopefully serve the majority of users needs that just use the defaults. GitHub's provider has some pretty sophisticated path rewriting. We might want to improve this in the future, but it should be sufficient for current needs in the time frame we have now.