Utility method to verify SSO Tokens (#1310)
This change introduces a utility method (`TokenExtensionMethods.HasValidToken`) which helps determine whether or not an SSO token has a valid, non-expired token in the cache. The check does not prompt users to log in if the cache is missing or the token is expired.

This utility method will be used in an upcoming change.