(endpoints 2.0) Update AWSSDK; stop injecting endpoints into AWSSDK (#1313)
This change updates the Toolkit to use the latest version of the AWS SDK. This is the first version of the SDK that uses endpoints 2.0, which takes a rule based approach to producing service endpoints for each service client. The Toolkit no longer injects the endpoints manifest into the SDK, since it no longer consumes this data.

A limited testing pass was made with this Toolkit, trying to use it in the me-central-1 region, which was known to have auth issues in the past when code was not properly configured. The functionality tested worked as expected: Lambda, S3, DynamoDB, DynamoDB Local, SQS, SNS, Publish to AWS, Legacy Publish (Beanstalk, Fargate, ECR), CloudFormation.