Store PAT in credentials manager; Remove PAT from git remote url (#1319)
This change updates how CodeCatalyst repos are cloned. The PAT is no longer stored as part of the repo's remote url. Instead, the PAT is stored in the Windows credential manager.

To support this change, `GitCredentials` and `NativeMethods` have been moved as-is to a more central project. This way, both CodeCommit and CodeCatalyst features can work with the same code responsible for saving credentials into the Windows credential manager.