Suppress post-clone failures; remove broken RemoteCodeContainer instantiation (#1333)
This change fixes a problem that was dormant until #1319. After the Toolkit clones a repo, it was causing an exception when trying to produce a payload containing details about the repo. RemoteCodeContainer cannot be instantiated when there is no browsable url. This was preventing the Toolkit from passing the cloned repo details to Visual Studio, which resulted in inconsistent behavior in the IDE after the repo was cloned.

To identify this case of problem in the future, a debug assert has been added, which will tell us that we have a problem during Toolkit development.

Co-developed with Shruti