Adjust CodeCatalyst validations to directly use SystemInfoTextBrushKey (#1334)
This change fixes a problem where the validation messages were not properly themed in VS 2019 - in dark mode the errors were using light text on light yellow background. The conceptual color was changed from ToolTipTextBrushKey to InfoTextBrushKey.

Over the course of this feature branch, we started to use `EnvironmentColors.SystemInfoTextBrushKey` for the first time in the repo. The problem is that our abstraction (`ToolkitThemes.InfoTextBrushKey`) does not get the proper value from the VS SDK. This could be an error within the VS SDK, since our other theme keys have not shown this behavior.

As a stopgap, places that were intended to use the Info Text color have been updated to directly use the VS SDK's EnvironmentColors.SystemInfoTextBrushKey. ValidationErrorsBox was moved into one of the VSSDK accessible projects, since this is the only place it is currently used.

Longer term, we'll want to revisit our validation presentation - we might want to consider a more canonical "red text on unchanged (transparent) background", closer matching VS own validation text presentation.