Reduce file watching scope when dragging and dropping from S3 Bucket browser to Windows explorer (#1341)
This change reduces the file watching scope and handling associated with dragging and dropping S3 objects from the S3 browser.

Previously, the Toolkit would watch all local drives and folders for the creation of a file `s3MaKeys.aws`. Whenever the Toolkit detects the creation of this file, it would interpret the contents of the file, and (if valid), copy specified S3 objects to a local folder. As long as the file referenced a valid credentials profile, and indicated the Toolkit instance's session Id (each session is assigned a random guid), the file would be processed. The file watcher would be started after the first S3 bucket browser is opened, and the watcher would run until the IDE is closed.

Now, the Toolkit will only create a file watcher whenever the user performs a drag and drop operation from the S3 bucket. Each drag and drop operation is assigned a unique Id (a random guid), and the watcher only responds to files associated with this operation Id. The watcher is stopped when the drag and drop operation completes. Additionally, the file being watched for only contains a reference to the operation Id -- the associated list of files to be downloaded from S3 is kept in memory. This means a hand-crafted file would be unable to instruct the Toolkit to download other files unexpectedly.

I initially set out to fully replace the file watcher, and have the Toolkit directly download files to the folder of interest. It turns out that the Windows drag and drop system does not support this use case very nicely:
- since Windows manages the drop location (a windows folder), we aren't able to determine where users are dropping their selected files. The file watcher approach that was already in use is a cheap way of achieving this (with the caveat that it only works for local drives). Google searches appear to arrive at the same conclusion.
- we could implement an "async drag and drop" system, but we would have to download the S3 objects to a temporary location, which would then become the "dropped" files once the user completes a drag and drop gesture. This would get complicated trying to ensure files aren't left lingering in a temporary location.

To make this change:
- commentary about the drag and drop approach was added to S3DragAndDropHandler
- S3DataObject.cs was moved to a folder containing all of the new drag and drop related code
- logic that was spread between ManifestWatcher.cs and S3DataObject.cs (relating to the manifest file) have been modelled in S3DragDropManifest. RIP ManifestWatcher.cs. S3DataObject.cs is now leaner, since the handling has been moved to new classes.