Decouple IViewModel and AccountViewModel from viewing Beanstalk Environments (#1266)
This change removes AWS Explorer related ViewModel (and AccountViewModel) coupling from the code that is used to view Beanstalk Environment resources.

The motivations for this change are:
- decouple AWS Explorer from plugin code to improve testability and performance (this is a longer, multi-change goal)
- isolate viewing Beanstalk Environment code so that it can be used from places like Publish to AWS
  - this will be done in an upcoming change
  - this will fix a problem where deployments made to regions other than what is selected the AWS Explorer can fail

To support this change, I introduced a model to represent a beanstalk environment that is agnostic of the AWS SDK (BeanstalkEnvironmentModel). This model becomes the input for commands related to viewing environments. Next, I replaced the AccountViewModel code with the newer credentials system (AwsConnectionSettings). These changes are able to provide the existing code with the same properties without changing functional implementations.

While more code improvements could be made in the affected Controller classes, I have limited the changes to what is necessary for isolating the "view environment" code.