Emit metric when the Toolkit provides Visual Studio with CodeArtifact package repo credentials (#1343)
We have metrics to understand when users configure the Toolkit to provide Visual Studio with a way to connect to CodeArtifact repos. This is typically a one time setup. We don't have metrics to understand how often those same instances of Visual Studio leverage a connection to CodeArtifact going forward. An example would be when Visual Studio fetches NuGet package 
references from CodeArtifact.

This change adds metrics (`codeartifact_credentialsRequest`) so that we better learn how often the Toolkit is helping establish Visual Studio to CodeArtifact connections.

The product's telemetry definitions were also updated to remove the `userId` type, since this is now defined in the toolkit-common repo.