Add DialogResult binding for clone dialog view models and move submit/cancel commands into view model. (#1351)
This change makes Window.DialogResult effectively bindable (via adding a behavior) to remove code from the code-behind and allow the view model to control dialog cancel/submit without having to have a reference to the view. It also moves the cancel/submit commands into the view model to remove additional code-behind code as well as keep the code within the view model to keep related code close to where it is used.

Added tests for DialogResult in view model as well as fixed a test for LocalPath makes CanExecute for submit true. The test didn't have a precondition to ensure CanExecute was initially false and used RelayCommand(null) which means that CanExecute always returns true. While the test was invalid, the actual dialog code did work as expected.

Verified that both Code Commit and Code Catalyst dialog's controls (particularly X to close window and Clone/Cancel buttons) behave as before and themes look correct in both VS2022 and VS2019.