Update Build tasks from .NET 4.5 to 4.7.2; add telemetry package reference updater build task (#1354)
This change introduces a new build task `update-telemetry` which automates updating all telemetry project references to the latest version. The build task was copied from `UpdateAwsSdkVersionsTask`, so while it lacks some finesse, it gets the job done.

I updated the build task to target .NET Framework 4.7.2 (up from 4.5). The motivation was to work with the nuget package reference I added, however this change should be okay in general, since 4.7.2 is currently the lowest .NET Framework version that is still in support.