Improve stability of LogExceptionAndForget tests (#1362)
This change addresses a problem where some LogExceptionAndForget tests fail in CI on occasion. This is caused because the tests were expecting failure details to be logged once a (failed) task completes, however LogExceptionAndForget does not wait for the logging to complete. As a bonus, these tests run faster, since they no longer employ a sleep-check loop.