General prompt and execute command (#1379)
Based on the EC2 commands that have been instrumented recently, a common pattern has emerged, that can be used in other plugins. This change is an attempt as creating common prompt and handler interfaces/classes.

To illustrate using the new classes, I phased out the SelectedElasticIpCommand and ElasticIpCommand class, converting their implementations over to the new handler interface. BaseEc2Command could be removed (with its implementations converted over), but this could take place as a separate change.