Fix STS Client creation; Add type checking to service client creation (#1382)
When publishing Lambda and Serverless projects, the Toolkit has been silently failing to create STS clients since #1275. The root cause is that our service client instantiation uses reflection, and requires class types instead of interfaces to be provided for creation.

This change fixes the two incorrect service client creation calls, adds validation into the client creation methods, and shows a debug assertion if we're debugging the Toolkit (so that we see this error during development).