Add CodeCatalyst icon to 'Clone a repository' dialog provider list (#1389)
Adds a CodeCatalyst icon to the 'Clone a repository' dialog provider list. The icon must be a png image in the package project set to Build Action = Resource. The Monikers.imagemanifest file is generated and must be set to a Build Action = Content and Include in VSIX = true. This file is shared between the VS 2019 and VS 2020 package projects and imports a Symbols.imagemanifest (same properties as the other manifest) to provide the actual resource path due to differing assembly names. The missing component from previous efforts is that the ProvideBindingPathAttribute must be set on the package to ensure that the package assembly is probed when looking for images.