Adjust CloudFormation projects to target .NET Framework 4.7.2 instead of 4.0 (#1390)
This change updates CloudFormation template projects to target .NET Framework 4.7.2 instead of 4.0. This will allow users to load projects on systems that do not have a .NET Framework 4.0 target pack installed. You're less likely to have that installed if you never worked with a major version of Visual Studio earlier than VS 2022.

The fact that it targets any framework is an artifact of how custom projects are implemented in Visual Studio extensions. We move away from targeting 4.0, because that version is no longer in support. Choosing 4.7.2 is safe, as it is the minimum version of .NET Framework supported by VS 2019 and 2022 (our current Toolkit hosts).

This addresses https://github.com/aws/aws-toolkit-visual-studio/issues/227