Merge publish feature to main (.NET 6; deploy tool 1.10.4) (#1394)
* Add check for .NET 6 when starting Publish to AWS from VS 2019 (#1346)

An upcoming change to the deploy tool will require NET 6 instead of .NET Core 3.1 because .NET Core 3.1 soon goes out of support. Publish to AWS is supported in VS 2019 and 2022. VS 2019 supports up to .NET Core 3.1, and VS 2022 includes .NET 6 support.

This change adds a check to the VS 2019 Toolkit. When starting Publish to AWS, the Toolkit will attempt to determine what version of .NET is available (by calling `dotnet --version`). 
- If .NET 6 or newer is detected, Publish to AWS loads normally
- If the installed version is too low, users are informed that they need to install .NET 6 or newer
- If there was a problem trying to detect the .NET version, users are informed that they should install .NET 6 (or newer) but are given the option to proceed, in case the Toolkit's detection is in error, and the user is able to successfully work with Publish to AWS

The deploy tool hasn't been updated yet, but will be soon. This change has been introduced into a feature branch, so that it is gated until we have this new deploy tool version.

* Update deploy tool to 1.10.4 (#1392)

This change updates the referenced deploy tool from 1.6.4 to 1.10.4. The biggest change with this update is the requirement to have .NET 6 (or newer) installed. The corresponding change to check for .NET 6 from the Toolkit was made in #1346.

This change was tested by running the integration test suite, and performing some deployments from the Toolkit.