Update SSO Credentials handling to accommodate token profiles (#1391)
Over the course of 2022, the credentials spec relating to SSO changed in order to accommodate token support for services like CodeCatalyst. The VS Toolkit was not updated to match the spec changes. This resulted in scenarios like https://github.com/aws/aws-toolkit-visual-studio/issues/313 , where SSO credentials profiles generated in the AWS CLI are not considered valid by the VS Toolkit. This change updates the Toolkit to align with the spec, and so that CLI produced profiles work in the Toolkit.

Spec changes:
- If either the sso_account_id or sso_role_name configurations values are present the profile MUST be resolved by the SSO credential provider.
  - The toolkit was not doing this before.
- If the profile and the sso-session both contain sso_region or sso_start_url then SDKs MUST validate that the values in both the profile and sso-session match. If either the sso_region or sso_start_url differ between the profile and sso-session, then SDK MUST raise a validation error.
  - The Toolkit used to consider it an error if a profile contained both sso_session and (sso_region or sso_start_url)
- If the sso_session configuration value is present, the profile MUST be resolved by the SSO token provider.
  - This is now secondary to the first spec above, allowing SSO based credentials to reference sso-session profiles