Fix weird state where connection message and resource tree appears together (#1437)
This change fixes some bugs/issues with the toolkit's credential-aws explorer environment.
In particular, addresses this issue where multiple SSO prompts would appear when connection expired and explorer/vs window was left idle.
A potential root cause for it could be file system checks running on a developer's system that touches the credentials file, triggering the toolkit's credential file watchers which would then attempt to create service clients while the connection is invalid.
The creation of service clients occurred due to a WPF binding that lazily loaded service nodes for the current connection.
This fix prevents the UX component representing TreeView of services to appear when the explorer has an invalid connection.

Bonus: The refresh icon action present on the AWS Explorer has been fixed to perform the same action as Retry link shown with an invalid connection (which attempts to refresh the current connection). Previously, it tried to recreate the tree with the current connection but did not always refresh the connection itself.