BuildTasks: Replace LitJson dependency for the one built into AWSSDK.Core (#1438)
We have some older infrastructure that leverages this repo's build tasks. As we update some of that infrastructure, we're seeing compatibility issues. This change removes the build failure relating to the `litjson` package not being strongly signed. The workaround is to use the version that is compiled into the AWS SDK (AWSSDK.Core).

We have plans to retire some of the affected build tasks, so this is mainly a stopgap.