Remove thirdparty/dotnet-lambda.dll (#1440)
`thirdparty/dotnet-lambda.dll` is bundled into the extension, but is not used. This change removes it.

It appears to have been added in 810bb38a6, but I didn't see how it was used.
It was phased out in 15155aa604, when the dotnet cli tooling code started copied into the repo under the `AmazonCLIExtensions` project.