Bring ARM Preview support into main branch (#1475)
This change merges the feature branch containing Preview support for Arm64 Visual Studio into the main branch.

With this change:
- the Toolkit can be installed into Arm versions of Visual Studio
- the Toolkit has a dismissible top level banner stating that Arm support is in preview
- the AWS Explorer has a fixed banner stating that Arm support is in preview
- the banner has gestures for users to send feedback and open GitHub issues
- the banner is only shown on ARM processors (not x64)
- telemetry now emits the os architecture (eg: x64, arm)

---

* Update VSIX mainifest to have ARM VS2022 installation target (#1342)

* Emit OS architecture (x86/ARM) with metrics (#1428)

This change updates the Toolkit's telemetry so that an `osArchitecture` field is emitted with each payload. The Toolkit detects the OS architecture at startup.

* Add "ARM Preview" InfoBar (without gestures) to main window (#1446)

This change adds an infobar to the main IDE window when the Toolkit is running on ARM systems. The infobar indicates that support for ARM is in Preview. When the Toolkit is run from an x64 system, the infobar is not displayed.

An upcoming change will add gestures to the infobar.

The code is mostly copy and pasted from the Telemetry notice code. This code is intended to be removed from the codebase after we exit the Preview phase, so I didn't do any refactoring.

I verified that the infobar displays when run on the arm dev kit, and that it does not display when run from my (x64) laptop.

* Add commands to ARM Preview Banner: Share feedback, File issue, Don't show again (#1448)

This change adds gestures to the ARM Preview notification banner.

- Share feedback: Sends feedback watermarked with the ARM Preview
- File issue: Opens a new GitHub issue, pre-filled with details to create a bug report
- Don't show again: Marks the banner to not be shown again, and closes the notification

* Add fixed ARM Preview banner into the AWS Explorer (#1450)

This change adds a dedicated "ARM Preview" banner into the AWS Explorer. This banner cannot be dismissed or removed, and only displays for users running on ARM based systems. The banner will remain until we release a version of the Toolkit where ARM support is GA.

The banner is deliberately not an infobar, and has been given a deliberate color scheme, so that it has visual differentiation from infobars, which can be seen as more disposable messaging. The coloring has been hardcoded, as there was no appropriate color definitions to fit the UX design. While this isn't ideal, this isn't intended to be used for a long period of time, and won't be seen by a large number of users, so we are accepting this antipattern.

To support this change, two of the arm preview commands were moved to a more central project (AWSToolkit) so that they can be referenced by the AWS Explorer.

* Only show ARM Preview banner in AWS Explorer from ARM systems (#1465)

This change limits the ARM Preview banner shown in the AWS Explorer (see #1450) so that it does not appear on x64 systems.

`ArmPreviewNotice` was moved to the AWSToolkit project so that it can be used by NavigatorViewModel

* Update the ARM Preview changelog entry (#1467)

This change updates the changelog to reflect the ARM Preview support.

---------

Co-authored-by: Norm Johanson <normj@amazon.com>
Co-authored-by: aws-toolkit-automation <43144436+aws-toolkit-automation@users.noreply.github.com>