Start of CodeWhisperer Settings (#1540)
This change defines a CodeWhisperer settings class, and integrates it into the Visual Studio options dialog ("Tools" > "Options").
The first setting (which will be used immediately) allows us to define a local location of the language server.

This change uses the [Community Toolkit](https://github.com/VsixCommunity/Community.VisualStudio.Toolkit) abstraction around Visual Studio settings - https://www.vsixcookbook.com/recipes/settings-and-options.html

This change updates certain VS 2022 projects to target .NET 4.8. The main AwsToolkitPackage.v17 project already targets 4.8, and Community Toolkit requires 4.8. It was an oversight that the other projects had not been updated yet.

SettingsText is a convenience class that consolidates the hardcoded text that is duplicated in the "Provide*" attributes of the toolkit package.