Start of CodeWhisperer language client implementation (#1543)
This change contains the start of the CodeWhisperer language client implementation. Additional future language clients have been kept in mind. The common code lives in ToolkitLspClient, and CodeWhispererClient contains handling specific to this first LSP integration.

Many portions of the client have been earmarked with TODO comments, which will be handled in upcoming work.

The language server has been set up to be associated with C# files. The server starts up and performs the initialization handshake, but does not do anything else. Future work is required before the language server is seen doing anything in the IDE:
- Toolkit needs to provide bearer tokens to the server
- Toolkit needs to make calls to generate code recommendations, then surface the responses in the IDE