Set up a project to place CodeWhisperer tests (#1554)
We need a place where CodeWhisperer tests (non-integration tests) can live. This change creates the test project, and seeds it with the first testable component.

Unfortunately there is a problem with the third party dependency that will be used by that first testable component, so this change stubs out the actual package reference (commented with a TODO), and will be a follow up change. This gets the boilerplate changes to support the test project out of the way.

I didn't use one of the existing VS test projects:
- CodeWhisperer is supported in VS 2022 (not VS 2019). This avoids the need to ifdef swaths of test code.
- these projects already spider out and take a dependency most of the repo in order to compile and run. This allows us to test the CodeWhisperer project's code faster.