Add jose-jwt reference; Strong-sign jose-jwt as part of the build. (#1557)
This change adds a reference to the `jose-jwt` package, which will be used to send encrypted credentials to language servers.

We have few good options available for securely encrypting JWT in .NET Framework. A complication of this package is that it is not strong-signed. With .NET, your solution has to either use fully strong-signed assemblies, or not at all, and this Toolkit uses them in order to ensure Visual Studio is loading the intended assemblies at run-time for this extension. We've taken the approach of automatically [self-strong-signing the package assembly](https://stackoverflow.com/questions/331520/how-to-fix-referenced-assembly-does-not-have-a-strong-name-error/331555#331555) as part of the build process. Details have been added to CONTRIBUTING.md explaining how this happens, and how to maintain it.

This process has been tested in both msbuild and Visual Studio.