Set and clear credentials from language server (#1563)
Passing credentials to a language service requires encrypted tokens and custom protocol messages.

This change introduces a new abstraction `IToolkitLspCredentials` which can be called (in future changes) to encrypt and transmit credentials over to the language server. This abstraction is composed of a component that handles the credentials encryption for the IDE session (`CredentialsEncryption`) and the JSON RPC proxy introduced in https://github.com/aws/aws-toolkit-visual-studio-staging/pull/1552 to call custom credentials protocol messages.