Update release script (#1565)
Dependabot is trying to update a python dependency we use in our release script (https://github.com/aws/aws-toolkit-visual-studio-staging/pull/1564). I tried using the release script with the updated dependency, and failures occur.

This change updates the release script in tandem with the dependency bump:

- Bump gitpython from 3.1.18 to 3.1.34 (addresses vulnerability concerns)
- use workaround from https://github.com/gitpython-developers/GitPython/issues/1630 so that the script can reset branches without failure. This problem was introduced in v3.1.32 of the dependency
- look for VS 2022 shell instead of VS 2019. We run this from VS 2022 now

I was able to verify that this script produces a release candidate branch, and pushes it to the repo. When we perform our next release, we will have to keep an eye out in case this change does not work as expected. If that happens, the mitigation will be to revert this change in order to unblock the release.