Add build action to produce a changelog that includes the currently queued changes (#1402)
This change adds a changelog to the build artifacts of the job that produces test vsix builds.
This allows us to see the changes queued for the next release at a glance.

The motivation for this change was that I routinely run these commands locally, typically more often as we get closer to a release. Then I have to revert all of the files that are changed, since it compiles the queued entries into the main changelog, then deletes those entry files. Now we have a way to grab these on the go.

To support this change, an additional flag was added to the changelog build task - FailIfNoChanges. By default, the script will continue to fail when there are no changes queued. Places like this GitHub Action can now suppress failures, since this action will run on all commits, and after we release, there generally aren't changelogs queued up immediately.