Set up stub CodeWhisperer margin (#1568)
This change sets up the initial CodeWhisperer Margin control. The margin control is where users interact with CodeWhisperer features, and resides in the margin below the code document. Clicking on the margin control (left or right mouse buttons) brings up a context menu with the operations that can be performed.

There is a lot that is stubbed in with this change:
- the CodeWhisperer image is a placeholder
- the menu commands do nothing

Some of the future tasks that will build from this change:
- set up margin control and menu item styling
- show connection status in the margin
- implement each menu item
- dynamically adjust the visibility of each menu item