ToolkitJoinableTaskFactoryProvider (#1571)
We should be using a JoinableTaskFactory in MEF components for switching between UI/background threads. By directly using things like `ThreadHelper.JoinableTaskFactory` and `Task.Run` we risk locking the UI thread. The AsyncPackage (Toolkit) is created with its own JoinableTaskFactory, but we don't have access to this from MEF components, and MEF components are activated independently from AsyncPackages.

We are able to create a JoinableTaskFactory from the same TaskContext used by Visual Studio. This change introduces a provider (ToolkitJoinableTaskFactoryProvider) that is capable of making a JoinableTaskFactory that we can import into MEF components.

This change also updates one MEF component (InvitationSection) to use ToolkitJoinableTaskFactoryProvider instead of using ThreadHelper. There are additional MEF components that can be migrated over time to use ToolkitJoinableTaskFactoryProvider, but updating one place is enough to let us know that ToolkitJoinableTaskFactoryProvider is working.

The motivation for making this change is to use it in some upcoming work that will create new MEF components.