Foundational components for CodeWhisperer Integration (#1575)
The CodeWhisperer integration has many different components. We want to provide a clean way for these components to interact with each other without excessive coupling. We also want the component code to be easily testable.

This change presents a skeleton implementation of the components we will implement first. The intention is to use MEF in order to create and inject the components together. An writeup has been added to the contributing guide to explain MEF, and how to troubleshoot it.

This design revolves around a central facade `CodeWhispererManager`, which is composed of smaller domain components. Larger IDE components import CodeWhispererManager, and use it to perform CodeWhisperer related operations. Most of the smaller domain components have been stubbed out, only a little amount of logic has been implemented for illustrative purposes.

SignInCommand was updated to illustrate how this overall design gets leveraged into something functional.

Most of the components is represented with an interface. This separation gives us ways to exercise each component in isolation. "Fake" versions of most components have been created, which also shows how we can test the components without redefining mocks in each test class.