Fix test that started failing (#1580)
This afternoon our test suite started failing on one of the "local endpoint" tests. This change makes the test pass again. I believe a change was made on the service endpoint.

This test (and others like it) should live in the integration test collection, but that work will have to take place another time.