Utility function to enable lsp tracing (#1579)
This change introduces code that we can use to help diagnose unexpected problems when communicating with a language server. When functionality isn't working, it can be caused when messages are not shaped the way we are expecting. Producing a raw log of the LSP channel can help with this.

We only ever want to do this during development! The helper code has been added to debug builds only, and is commented out by default.