Add projects for Common Test code involving VSSDK (#1584)
This change defines a project (per VS version - v16 and v17) that contains code that is common to tests that require access to the VS SDK.
Until now, this code has been defined in a shared project `AWSToolkit.Tests.Common.VS.Shared`, and test projects took a direct dependency on this shared project. This doesn't scale nicely, because when new code that requires a NuGet dependency is added to the shared project, each of the referring projects have to be updated.

To clean this up, "concrete" projects have been created for each VS version that reference the shared project. Test projects that had a reference to the shared project have been updated to reference these "concrete" projects instead. This gives us one set of projects (one per VS version) to update whenever new code is added to AWSToolkit.Tests.Common.VS.Shared.

This change is being made in advance of a change that will add code to AWSToolkit.Tests.Common.VS.Shared, and introduce new dependencies.