Add support for testing code that uses JoinableTaskFactory (#1587)
This change sets up supporting test code and dependencies for testing code that uses the JoinableTaskFactory and/or the VSSDK.

The change is based on guidance in https://github.com/Microsoft/vs-threading/blob/main/doc/testing_vs.md and https://github.com/microsoft/vssdktestfx/blob/main/src/Microsoft.VisualStudio.Sdk.TestFramework.Xunit/README.md . Basically we use a Fixture, which stubs in a selection of central VS SDK objects. One of the things that this sets up is code that ThreadHelper requires, which means we should be able to validate UI/background thread use in tests.

This change only supports VS 2022 and newer. The plan is to use it in CodeWhisperer work first, which is also only supporting VS 2022 and newer.