Toolkit can process suggestion responses from language server (#1581)
This change is a continuation of the work in #1577 to load suggestions from the language server. In this change, the Toolkit is able to interpret the inline completions returned, and convert them to a Toolkit model.

We don't have suggestions displaying in the IDE at this time. Until then, the suggestions are written to the output pane as a way to show that we have the suggestions available within the Toolkit. I also wrapped the suggestion request in a Task Notifier, so that we have visual feedback while we're developing this integration.