Add cleanup logic for cached LSP versions (#1610)
This adds cleanup logic to remove old versions of the CW language server in the cache location. Cleanup helps keep the storage reasonably clean.
* It is executed on a background thread asynchronously and is triggered after a 10sec delay after the Language server download process completes successfully.
* It cleans up versions cached that are in the toolkit compatible version range. This includes:
       * deleting any delisted versions in that range
       * deleting all but the highest 2 versions in that range

It also introduces handling for de-listed versions by the toolkit. A version is considered de-listed if it is not present in/removed from the manifest, or is marked with a "de-listed" flag in the manifest. Handling is described as followed:
* When Toolkit look for the highest compatible version of the language server, it excludes all de-listed versions from consideration.
* Toolkit deletes any delisted versions if found during cleanup process