Push CodeWhisperer settings to language server on Init and on Save (#1606)
This change pushes CodeWhisperer settings to the language server whenever the language client completes the Initialization handshake, and whenever CodeWhisperer settings are saved.

The approach taken has been generalized to support future language servers.

The way it works is to make the base language client raise an event when the Initialization handshake is successful. We then define a separate component (SettingsPublisher) that watches for this event (as well as the Settings Saved event), which sends the LSP protocol message to the language server indicating that the workspace configuration has changed.

A follow-up change will add support for the language server to request the current configuration state from the client. Some of the events have been roughed in as part of this change.